refactor: moved transaction columns to the correct place
